Better organise library file structure.
