+++ examples/v2/lab_tests.py
@@ -4,7 +4,7 @@ config.setup_examples()
 import infermedica_api
 
 if __name__ == '__main__':
-    api = infermedica_api.get_api('v2')
+    api: infermedica_api.ModelAPIv2Connector = infermedica_api.get_api('v2')
 
     print('Laboratory tests list:')
     print(api.lab_test_list(), end="\n\n")
