+++ examples/v2/triage.py
@@ -1,13 +1,14 @@
+from infermedica_api.connectors.v2.models import Diagnosis
 from .. import config
 
 config.setup_examples()
 import infermedica_api
 
 if __name__ == '__main__':
-    api = infermedica_api.get_api('v2')
+    api: infermedica_api.ModelAPIv2Connector = infermedica_api.get_api('v2')
 
     # Prepare the diagnosis request object
-    request = infermedica_api.Diagnosis(sex='male', age=30)
+    request = Diagnosis(sex='male', age=30)
     request.add_symptom('s_1193', 'present')
     request.add_symptom('s_488', 'present')
     request.add_symptom('s_418', 'absent')
