+++ infermedica_api/connectors/v2/__init__.py
@@ -1,0 +1,3 @@
+from .basic import BasicAPIv2Connector
+from .standard import APIv2Connector
+from .model import ModelAPIv2Connector
