+++ infermedica_api/connectors/v2/models/base.py
