+++ infermedica_api/connectors/v2/models/condition.py
