+++ infermedica_api/connectors/v2/models/red_flag.py
