+++ infermedica_api/connectors/v2/models/symptom.py
