+++ infermedica_api/connectors/v3/__init__.py
@@ -1,0 +1,2 @@
+from .basic import BasicAPIv3Connector
+from .standard import APIv3Connector, ConceptType
